+++ crates/biome_js_analyze/tests/specs/nursery/noDefaultExport/valid.cjs
@@ -1,0 +1,3 @@
+module.exports = class {};
+exports = function () {};
+module.exports.A = class A {};