+++ crates/biome_js_syntax/src/identifier_ext.rs
@@ -1,8 +1,8 @@
 use crate::{
-    AnyJsName, JsIdentifierAssignment, JsLiteralExportName, JsReferenceIdentifier, JsSyntaxToken,
-    JsxReferenceIdentifier,
+    inner_string_text, AnyJsName, JsIdentifierAssignment, JsLiteralExportName,
+    JsReferenceIdentifier, JsSyntaxToken, JsxReferenceIdentifier,
 };
-use biome_rowan::{declare_node_union, SyntaxResult};
+use biome_rowan::{declare_node_union, SyntaxResult, TokenText};
 
 declare_node_union! {
     pub AnyJsIdentifierUsage = JsReferenceIdentifier | JsIdentifierAssignment | JsxReferenceIdentifier
@@ -19,8 +19,36 @@ impl AnyJsIdentifierUsage {
 }
 
 impl JsLiteralExportName {
-    pub fn is_default(&self) -> SyntaxResult<bool> {
-        Ok(self.value()?.text_trimmed() == "default")
+    /// get the exported name, stripping the quotes if it is a string.
+    ///
+    /// ```
+    /// use biome_js_factory::make;
+    /// use biome_rowan::TriviaPieceKind;
+    ///
+    /// let export_name = make::js_literal_export_name(make::js_string_literal("foo")
+    ///     .with_leading_trivia(vec![(TriviaPieceKind::Whitespace, " ")]));
+    /// assert_eq!(export_name.inner_string_text().unwrap().text(), "foo");
+    /// ```
+    pub fn inner_string_text(&self) -> SyntaxResult<TokenText> {
+        Ok(inner_string_text(&self.value()?))
+    }
+
+    /// Returns `true` if the export name is `default`.
+    ///
+    /// ```
+    /// use biome_js_factory::make;
+    /// use biome_rowan::TriviaPieceKind;
+    ///
+    /// let export_name = make::js_literal_export_name(make::js_string_literal("default"));
+    /// assert!(export_name.is_default());
+    ///
+    /// let export_name = make::js_literal_export_name(make::js_string_literal("foo"));
+    /// assert!(!export_name.is_default());
+    /// ```
+    pub fn is_default(&self) -> bool {
+        self.inner_string_text()
+            .map(|x| x.text() == "default")
+            .unwrap_or(false)
     }
 }